tidy
git-svn-id: http://svn.bramp.net:8080/svn/PhD/Work/userspace/threadnetperf@2058 e02701c3-087c-644a-a2a3-551e05f81084

+++ print.c
@@ -69,7 +69,7 @@ void print_results( const struct settings *settings, const struct stats *stats )
 	pthread_mutex_unlock( &printf_mutex );
 }
 
-void print_stats(float sum, float sumsquare, float mean, float variance) {
+void print_stats(double sum, double sumsquare, double mean, double variance) {
 	printf("sum %f sumsquare %f mean %f variance %.f\n",sum, sumsquare, mean, variance);
 }
